+++ app/src/main/java/kozlov/artyom/avitoweather/domain/entity/WeatherHourly.kt
@@ -3,5 +3,5 @@ package kozlov.artyom.avitoweather.domain.entity
 data class WeatherHourly(
     val hour: String,
     val picture: String,
-    val temp: Double
+    val temp: Int
 )
